convert all read to ts
